+++ src/global.rs
@@ -95,12 +95,6 @@ impl<'e> IntoLisp<'e> for &'e GlobalRef {
     }
 }
 
-impl<'e> IntoLisp<'e> for &'e OnceGlobalRef {
-    #[inline(always)]
-    fn into_lisp(self, env: &'e Env) -> Result<Value<'e>> {
-        Ok(self.bind(env))
-    }
-}
 
 impl<'e> Value<'e> {
     /// Creates a new [`GlobalRef`] for this value.
@@ -112,10 +106,50 @@ impl<'e> Value<'e> {
     }
 }
 
+/// Declares global references. These will be initialized when the module is loaded.
+#[doc(hidden)]
+#[macro_export]
+macro_rules! global_refs {
+    ($($name:ident)*) => {
+        $(
+            #[allow(non_upper_case_globals)]
+            pub static $name: &'static $crate::OnceGlobalRef = {
+                static X: $crate::OnceGlobalRef = $crate::OnceGlobalRef::new();
+                &X
+            };
+        )*
+    };
+    ($registrator_name:ident ($init_method:ident) =>
+        $(
+            $name:ident $( => $lisp_name:expr )?
+        )*
+    ) => {
+        $crate::global_refs! {
+            $($name)*
+        }
+
+        #[$crate::deps::ctor::ctor]
+        fn $registrator_name() {
+            $crate::init::__GLOBAL_REFS__.try_lock()
+                .expect("Failed to acquire a write lock on the list of initializers for global refs")
+                .push(::std::boxed::Box::new(|env| {
+                    $(
+                        #[allow(unused_variables)]
+                        let name = $crate::deps::emacs_macros::lisp_name!($name);
+                        $( let name = $lisp_name; )?
+                        $crate::OnceGlobalRef::$init_method(&$name, env, name)?;
+                    )*
+                    Ok(())
+                }));
+        }
+    };
+}
+
 /// A [`GlobalRef`] that can be initialized once. This is useful for long-lived values that should
-/// be initialized when the module is loaded, such as frequently-used symbols.
+/// be initialized when the dynamic module is loaded. A typical use case is specifying
+/// frequently-used symbols, which can be done with the help of the macro [`use_symbols!`].
 ///
-/// [`GlobalRef`]: struct.GlobalRef.html
+/// [`use_symbols`]: crate::use_symbols
 #[derive(Debug)]
 #[repr(transparent)]
 pub struct OnceGlobalRef {
@@ -123,31 +157,45 @@ pub struct OnceGlobalRef {
 }
 
 impl OnceGlobalRef {
-    pub(crate) const fn new() -> Self {
+    pub const fn new() -> Self {
         Self { inner: OnceCell::new() }
     }
 
     /// Initializes this global reference with the given function.
     #[doc(hidden)]
-    pub fn init<F: FnOnce(&Env) -> Result<Value>>(&self, env: &Env, f: F) -> Result<()> {
+    pub fn init<F: FnOnce(&Env) -> Result<Value>>(&self, env: &Env, f: F) -> Result<&GlobalRef> {
         let g = f(env)?.make_global_ref();
         self.inner.set(g).expect("Cannot initialize a global reference more than once");
-        Ok(())
+        Ok(self.inner.get().expect("Failed to get an initialized OnceGlobalRef"))
     }
 
     /// Points this global reference to an interned Lisp symbol with the given name.
+    ///
+    /// This should be called once, during module initialization.
     #[doc(hidden)]
-    pub fn init_to_symbol(&self, env: &Env, name: &str) -> Result<()> {
+    pub fn init_to_symbol(&self, env: &Env, name: &str) -> Result<&GlobalRef> {
         self.init(env, |env| env.intern(name))
     }
 
     /// Points this global reference to the function bound to the Lisp symbol with the given name.
+    ///
+    /// This should be called once, during module initialization.
+    ///
+    /// If the symbol is later bound to another function, this global reference will still point to
+    /// the old function. Therefore, this is best used for built-in and primitive functions.
     #[doc(hidden)]
-    pub fn init_to_function(&self, env: &Env, name: &str) -> Result<()> {
+    pub fn init_to_function(&self, env: &Env, name: &str) -> Result<&GlobalRef> {
         self.init(env, |env| {
             let symbol = env.intern(name)?;
             env.call("indirect-function", [symbol])
         })
+    }
+}
+
+impl<'e> IntoLisp<'e> for &'e OnceGlobalRef {
+    #[inline(always)]
+    fn into_lisp(self, env: &'e Env) -> Result<Value<'e>> {
+        Ok(self.bind(env))
     }
 }