+++ src/init.rs
@@ -10,7 +10,7 @@ use std::{
 
 use once_cell::sync::Lazy;
 
-use crate::{Env, Value, Result};
+use crate::{Env, Value, Result, ErrorKind};
 
 #[doc(hidden)]
 #[macro_export]
@@ -51,7 +51,17 @@ type FnMap = HashMap<String, InitFn>;
 /// [`emacs_module_init`].
 ///
 /// [`emacs_module_init`]: https://www.gnu.org/software/emacs/manual/html_node/elisp/Dynamic-Modules.html
-pub static __PRE_INIT__: Lazy<Mutex<Vec<InitFn>>> = Lazy::new(|| Mutex::new(vec![]));
+pub static __GLOBAL_REFS__: Lazy<Mutex<Vec<InitFn>>> = Lazy::new(|| Mutex::new(vec![]));
+
+/// Functions that will be called by [`emacs_module_init`] to define custom error signals.
+///
+/// They are called before loading module metadata, e.g. module name, function prefix.
+///
+/// This list is populated when the OS loads the dynamic library, before Emacs calls
+/// [`emacs_module_init`].
+///
+/// [`emacs_module_init`]: https://www.gnu.org/software/emacs/manual/html_node/elisp/Dynamic-Modules.html
+pub static __CUSTOM_ERRORS__: Lazy<Mutex<Vec<InitFn>>> = Lazy::new(|| Mutex::new(vec![]));
 
 /// Functions that will be called by [`emacs_module_init`] to define the module functions.
 ///
@@ -88,23 +98,36 @@ fn check_gc_bug_31238(env: &Env) -> Result<()> {
 }
 
 #[inline]
-pub fn initialize<F>(env: &Env, f: F) -> os::raw::c_int
+pub fn initialize<F>(env: &Env, init: F) -> os::raw::c_int
     where
         F: Fn(&Env) -> Result<Value<'_>> + panic::RefUnwindSafe,
 {
     let env = panic::AssertUnwindSafe(env);
     let result = panic::catch_unwind(|| match (|| {
-        for f in __PRE_INIT__.try_lock().expect("Failed to acquire a read lock on the list of initializers").iter() {
-            f(&env)?;
+        for init_global_ref in __GLOBAL_REFS__.try_lock()
+            .expect("Failed to acquire a read lock on the list of initializers for global-refs").iter() {
+            init_global_ref(&env)?;
         }
-        env.define_errors()?;
+        env.define_core_errors()?;
         check_gc_bug_31238(&env)?;
-        f(&env)
+        for define_error in __CUSTOM_ERRORS__.try_lock()
+            .expect("Failed to acquire a read lock on the list of initializers for custom error signals").iter() {
+            define_error(&env)?;
+        }
+        init(&env)
     })() {
         Ok(_) => 0,
         Err(e) => {
-            env.message(format!("Error during initialization: {:#?}", e))
-                .expect("Fail to message Emacs about error");
+            if let Some(ErrorKind::Signal { symbol, data }) = e.downcast_ref::<ErrorKind>() {
+                env.call("message", (
+                    "Error during initialization: symbol: %s data: %s",
+                    unsafe { symbol.value(&env) },
+                    unsafe { data.value(&env) },
+                ))
+            } else {
+                env.message(format!("Error during initialization: {:#?}", e))
+            }
+                .expect("Failed to message Emacs about initialization error");
             1
         }
     });
@@ -112,7 +135,7 @@ pub fn initialize<F>(env: &Env, f: F) -> os::raw::c_int
         Ok(v) => v,
         Err(e) => {
             env.message(format!("Panic during initialization: {:#?}", e))
-                .expect("Fail to message Emacs about panic");
+                .expect("Failed to message Emacs about initialization panic");
             2
         }
     }