+++ src/subr.rs
@@ -1,8 +1,6 @@
-#![allow(non_upper_case_globals)]
-
 global_refs! {common(init_to_function) =>
     cons car cdr
-    vector make_vector => "make-vector"
+    vector make_vector
     list
     message
 }